+++ src/aasm/__init__.py
@@ -1,2 +1,2 @@
-from .public_v50 import *
-from .public_v50 import __version__, AASMEngine, validate_public_api_contract, public_api_contract, PUBLIC_API_CONTRACT
+from .public_v51 import *
+from .public_v51 import __version__, AASMEngine, validate_public_api_contract, public_api_contract, PUBLIC_API_CONTRACT